refactor(slack): clean up host-beta wiring
